app과 ingress 분리 (#136)
ingress는 각 앱에서 직접 설정하도록 함
