+++ packages/pulumi/src/index.ts
@@ -1,4 +1,4 @@
+export { App } from './app';
 export { DopplerSecret } from './doppler-secret';
 export { IAMServiceAccount } from './iam-service-account';
 export { Redirect } from './redirect';
-export { Service } from './service';
